pyproject.toml
[project]
name = "drawbot-cli"
version = "0.1.0"
description = "Headless, skia-native DrawBot CLI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "pyyaml>=6.0",
    "skia-python",
    "fonttools[unicode]",
    "numpy",
    "uharfbuzz",
    "python-bidi==0.4.2",
    "unicodedata2",
    "skia-pathops",
    "blackrenderer",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
drawbot = "drawbot_cli.app:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["drawbot_cli"]

[tool.basedpyright]
extraPaths = ["."]
pythonVersion = "3.11"
